pyproject.toml
[project]
name = "raudio"
version = "0.1.0"
description = "Lance table for easytranscriber AudioMetadata JSON — chunk-centric schema with embedded PCM audio, word-level alignments, and Tantivy FTS via LanceDB."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyarrow>=21.0.0",
    "lancedb>=0.30.0",
    "msgspec>=0.21.1",
    "pylance>=4.0.0",
    "typer>=0.12.0",
    # FastAPI viewer backend.
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Transcription stack. Torch pinned to a cu128 build: driver 570.x supports
    # up to CUDA 12.8, and cu130 wheels fail to initialize ("driver too old").
    "easytranscriber",
    "torch==2.11.0+cu128",
    "torchaudio==2.11.0+cu128",
]

[project.scripts]
raudio = "raudio.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/raudio", "backend"]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu128" }]
torchaudio = [{ index = "pytorch-cu128" }]
